PostgreSQL移行

+++ stationapi/src/use_case/dto/company.rs
@@ -12,8 +12,8 @@ impl From<Company> for GrpcCompany {
             name_english_short: company.company_name_en,
             name_english_full: company.company_name_full_en,
             url: company.company_url,
-            r#type: company.company_type as i32,
-            status: company.e_status as i32,
+            r#type: company.company_type,
+            status: company.e_status,
         }
     }
 }
